Query WiFi info (#10)
* Query WiFi info

This queries SSID, frequency, bitrate, link quality and signal level
for every wireless device reported in /proc/net/wireless.

* Call getWifiData

Also had to include <sys/socket.h> to get it to compile

* Serialize wifi data to ROS message

* Check ioctl for return code of -1

man ioctl suggests that the success code is usually 0, but the erro
code must be -1. Let's check for the eror code instead of the possible
success code.

* Don't query other params if SSID is missing

* Move the driver socket to the SysStats

Instead of tying the lifetime of the query socket to WiFi instances,
we'll reuse the same socket for all WiFi devices and we'll keep it
alive as long as the program is alive.

* Move the iw_req structure from WiFi instances to queryWifi method

Since it can be reused for different devices, the lifetime matches
the queryWifi method, not the device info.

* If you can't query the SSID, don't calculate the other data

* Do not erase Wifi devices, but set their info to empty.